fix(calendar): demo拆解与规范
