+++ src/packages/calendar/types.ts
@@ -14,3 +14,7 @@ export interface CalendarMonthInfo {
 export type CalendarValue = string | string[]
 
 export type CalendarType = 'single' | 'range' | 'multiple' | 'week'
+
+export type CalendarRef = {
+  scrollToDate: (date: string) => void
+}
